[RENAME/#11] adapter 패키지로 옮김 & 파일명 수정
