+++ app/src/main/java/com/puzzling/puzzlingaos/presentation/home/mypage/MyPageFragment.kt
@@ -8,6 +8,9 @@ import com.puzzling.puzzlingaos.R
 import com.puzzling.puzzlingaos.base.BaseFragment
 import com.puzzling.puzzlingaos.data.model.response.ResponseMyPageProjectDto
 import com.puzzling.puzzlingaos.databinding.FragmentMyPageBinding
+import com.puzzling.puzzlingaos.presentation.home.mypage.adapter.MyProjectBottomAdapter
+import com.puzzling.puzzlingaos.presentation.home.mypage.adapter.MyProjectAdapter
+import com.puzzling.puzzlingaos.presentation.home.mypage.adapter.MyProjectTopAdapter
 
 class MyPageFragment : BaseFragment<FragmentMyPageBinding>(R.layout.fragment_my_page) {
 
@@ -28,7 +31,7 @@ class MyPageFragment : BaseFragment<FragmentMyPageBinding>(R.layout.fragment_my_
 
     private fun initAdapter() {
         val myProjectAdapter = MyProjectAdapter()
-        val concatAdapter = ConcatAdapter(TopAdapter("지니"), myProjectAdapter, BottomAdapter())
+        val concatAdapter = ConcatAdapter(MyProjectTopAdapter("지니"), myProjectAdapter, MyProjectBottomAdapter())
 
         with(binding) {
             rcvMyPageMain.adapter = concatAdapter
